File to detect the hand gesture finished
There's still a bug. If you don't make any sign in the first three
seconds, it will understand that as a "one". I tried to solve it
creating a boolean variable count_change that returns True if
the angle is <= 90. However, that made the algorithm less accurate
so I decided not to apply it.